Implement TLSAuthEnable
Tested dgx-a100-dp in qemu. Starting fresh image,
bmcweb is accessible through http 80 and doesn't require authentication, and the relevant paths being unavailable.
After issuing: curl -k -X POST -d  '{}' ${bmc}/redfish/v1/Managers/bmc/Oem/Nvidia/TLSAuth/Enable
bmcweb restarts itself, and uses https 443 and requires authentication for relevant features,
and the previously missing paths become available again.

Fixes jira https://jirasw.nvidia.com/projects/DGXOPENBMC/issues/DGXOPENBMC-3959

+++ http/http_server.hpp
@@ -104,31 +104,34 @@ class Server
     void loadCertificate()
     {
 #ifdef BMCWEB_ENABLE_SSL
-        namespace fs = std::filesystem;
-        // Cleanup older certificate file existing in the system
-        fs::path oldCert = "/home/root/server.pem";
-        if (fs::exists(oldCert))
+        if constexpr (std::is_same<Adaptor, boost::beast::ssl_stream<boost::asio::ip::tcp::socket>>::value)
         {
-            fs::remove("/home/root/server.pem");
+            namespace fs = std::filesystem;
+            // Cleanup older certificate file existing in the system
+            fs::path oldCert = "/home/root/server.pem";
+            if (fs::exists(oldCert))
+            {
+                fs::remove("/home/root/server.pem");
+            }
+            fs::path certPath = "/etc/ssl/certs/https/";
+            // if path does not exist create the path so that
+            // self signed certificate can be created in the
+            // path
+            if (!fs::exists(certPath))
+            {
+                fs::create_directories(certPath);
+            }
+            fs::path certFile = certPath / "server.pem";
+            BMCWEB_LOG_INFO << "Building SSL Context file=" << certFile;
+            std::string sslPemFile(certFile);
+            std::vector<char> &pwd = lsp::getLsp();
+            ensuressl::ensureOpensslKeyPresentEncryptedAndValid(
+                sslPemFile, &pwd, lsp::passwordCallback);
+            std::shared_ptr<boost::asio::ssl::context> sslContext =
+                ensuressl::getSslContext(sslPemFile);
+            adaptorCtx = sslContext;
+            handler->ssl(std::move(sslContext));
         }
-        fs::path certPath = "/etc/ssl/certs/https/";
-        // if path does not exist create the path so that
-        // self signed certificate can be created in the
-        // path
-        if (!fs::exists(certPath))
-        {
-            fs::create_directories(certPath);
-        }
-        fs::path certFile = certPath / "server.pem";
-        BMCWEB_LOG_INFO << "Building SSL Context file=" << certFile;
-        std::string sslPemFile(certFile);
-        std::vector<char> &pwd = lsp::getLsp();
-        ensuressl::ensureOpensslKeyPresentEncryptedAndValid(
-            sslPemFile, &pwd, lsp::passwordCallback);
-        std::shared_ptr<boost::asio::ssl::context> sslContext =
-            ensuressl::getSslContext(sslPemFile);
-        adaptorCtx = sslContext;
-        handler->ssl(std::move(sslContext));
 #endif
     }
